+++ db_data/ib_buffer_pool
@@ -1,22 +1,35 @@
 3,2
 3,1
 3,3
-10,105
-10,115
-10,181
-10,103
-10,177
-10,247
-10,161
-10,223
-10,2
+10,136
+10,304
+10,74
+10,108
+10,90
+10,296
+10,182
+11,14
+11,6
+11,12
+11,10
 9,7
-9,5
 9,2
 9,3
+8,12
+8,9
 8,15
-8,7
-8,5
+110,3
+110,2
+110,1
+110,0
+109,3
+109,2
+109,1
+109,0
+108,3
+108,2
+108,1
+108,0
 20,3
 20,2
 20,1
@@ -33,39 +46,25 @@
 11,26
 11,25
 11,24
-11,23
 11,22
-11,21
-11,20
-11,19
-11,18
 11,17
-11,16
-11,15
-11,14
 11,13
-11,12
 11,11
-11,10
-11,9
-11,8
 11,7
-11,6
-11,5
-11,4
 11,3
 11,2
 11,1
-11,0
-10,315
-10,314
-10,310
-10,304
-10,287
-10,282
-10,263
-10,259
-10,256
-10,248
+10,291
 18,1
-10,246
+10,282
+18,3
+10,271
+10,263
+10,236
+14,1
+10,232
+14,3
+10,201
+13,1
+10,194
+13,3
