+++ main.go
@@ -23,10 +23,8 @@ func main() {
 	mux.HandleFunc("/movie/:id", show)
 	mux.HandleFunc("/movie/create", create)
 	mux.HandleFunc("/movie/:id/update", update)
-	log.Printf("updateの後ろまで来たしん")
 	mux.HandleFunc("/movie/:id/delete", delete)
 	log.Printf("呼んでるしん")
-
 
 	http.ListenAndServe(":8000", mux)
 	log.Printf("listen終わったよ")
